% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_fips.R
\name{ave_hourly}
\alias{ave_hourly}
\title{Average across hourly stations}
\usage{
ave_hourly(stationdata)
}
\description{
Average across hourly stations
}
\examples{
\dontrun{
average_data <- ave_hourly(stationdata)
aug_ave <- with(average_data,
subset(average_data, average_data$date_time > as.POSIXct('1992-08-01 00:00:00') &
average_data$date_time < as.POSIXct('1992-08-31 00:00:00')))
ggplot(aug_ave, aes(x = date_time, y = mean)) + geom_line() + theme_minimal()
}
}

