% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_fips new.R
\name{average_weather}
\alias{average_weather}
\title{Average weather data across multiple stations}
\usage{
average_weather(weather_data, start_date = NULL, end_date = NULL)
}
\arguments{
\item{weather_data}{a \code{meteo_pull_monitors} data.frame}

\item{start_date}{a date in "yyyy-mm-dd" format - the earliest date you want
in your dataset.
\code{start_date}.}

\item{end_date}{a date in "yyyy-mm-dd" format - the lastest date you want in
your dataset.}
}
\description{
\code{average_weather} returns a data.frame with daily values for
precipitation, maximum and minimum temperature averaged across multiple
stations. The resulting data.frame has a column \code{n_reporing} showing the
number of stations contributing to the average values for each day. Values for
this column can range from 1 to the number of unique stations in your
\code{meteo_pull_monitors} data.frame.
}
\details{
If a station has missing data for either \code{prcp}, \code{tmax}, or
\code{tmin} on a particular day, it will not contribute to the average value
for any of those weather variables for that day.
}
\examples{
\dontrun{
monitors <- c("ASN00095063", "ASN00024025", "ASN00040112")
obs <- meteo_pull_monitors(monitors)
avg <- average_weather(obs, "1999-01-01", "2013-01-01")
}
}

