% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_outlines}
\alias{county_outlines}
\title{County cartographic boundary shapefiles}
\format{A "SpatialPolygonsDataFrame" object}
\source{
\url{https://www.census.gov/geo/maps-data/data/cbf/cbf_counties.html}
}
\usage{
county_outlines
}
\description{
A "SpatialPolygonsDataFrame" object. This shapefile downloaded from the U.S.
Census 2010 Cartographic Boundary Shapefiles for counties. The
gz_2010_us_050_00_500k zip file was downloaded from the website listed in
"Source."
}
\keyword{datasets}

