% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_fips.R
\name{county_timeseries}
\alias{county_timeseries}
\title{Write daily weather timeseries files for U.S. counties}
\usage{
county_timeseries(fips, percent_coverage, date_min, date_max, var,
  average_data = TRUE, out_directory, out_type = "rds")
}
\arguments{
\item{percent_coverage}{A numeric value in the range of 0 to 1 that specifies
the desired percentage coverage for the weather variable (i.e., what
percent of each weather variable must be non-missing to include data from
a monitor when calculating daily values averaged across monitors.
(Optional.)}

\item{date_min}{A character string giving the earliest date you want
   in your dataset in "yyyy-mm-dd" format. (Optional.)
\code{date_min}.}

\item{date_max}{A character string giving the latest date you want
in your dataset in "yyyy-mm-dd" format. (Optional.)}

\item{var}{A character vector specifying desired weather variables. For
example, var = c("TMIN", "TMAX", "PRCP"). (Optional.)}

\item{average_data}{TRUE / FALSE to indicate if you want the function to
average daily weather data across multiple monitors.}

\item{out_directory}{The absolute or relative pathname for the directory
where you would like the timeseries files to be saved.}

\item{out_type}{A character string indicating that you would like either .rds
files (out_type = "rds") or .csv files (out_type = ".csv"). This option
defaults to .rds files.}
}
\value{
Writes out a directory with daily weather files for each FIPS code
specified.
}
\description{
Given a vector of U.S. county FIPS codes, this function creates timeseries
dataframes giving: 1. the values for specified weather variables, and 2. the
number of weather stations contributing to the average for each day within the
specified date range.
}
\note{
If the function is unable to pull weather data for a particular county
given the specified percent coverage, date range, and/or weather variables,
\code{county_timeseries} will not produce a file for that county.
}
\examples{
\dontrun{
county_timeseries(fips = c("41005", "13089"), percent_coverage = 0.90,
           date_min = "2000-01-01", date_max = "2000-01-10",
           var = c("TMAX", "TMIN", "PRCP"),
           out_directory = "~/timeseries_data")
}

}

