% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_helpers.R
\name{daily_stationmap}
\alias{daily_stationmap}
\title{Plot daily weather stations for a particular county.}
\usage{
daily_stationmap(fips, daily_data, point_color = "firebrick",
  point_size = 2, station_label = FALSE)
}
\arguments{
\item{fips}{A five-digit FIPS county code.}

\item{daily_data}{A list returned from the function \code{daily_df}.}

\item{point_color}{point_color The specified \code{ggplot2} color for each point
representing the location of a station.}

\item{point_size}{The specified \code{ggplot2} size for each point
representing the location of a station.}

\item{station_label}{TRUE / FALSE If TRUE, includes labels giving the id for
monitor on the map. The default is FALSE.}
}
\value{
A plot showing points for all weather stations for a particular
   county satisfying the conditions present in \code{daily_df}'s
   arguments (date range and/or var). 2010 U.S. Census cartographic boundary
   shapefiles are used to provide county outlines.
}
\description{
This function produces a map with points indicating stations that contribute
to the weather data in the argument \code{daily_data}.
}
\examples{
\dontrun{
stations <- daily_stations(fips = "12086", date_min = "1992-08-01",
                          date_max = "1992-08-31")
daily_data <- daily_df(stations = stations, coverage = 0.90,
                      var = c("tmax", "tmin", "prcp"),
                      date_min = "1992-08-01", date_max = "1992-08-31")
daily_stationmap("12086", daily_data)
}

}

