% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_fips.R
\name{daily_timeseries}
\alias{daily_timeseries}
\title{Write daily weather timeseries files for U.S. counties}
\usage{
daily_timeseries(fips, coverage = NULL, date_min, date_max, var,
  average_data = TRUE, out_directory, out_type = "rds")
}
\arguments{
\item{fips}{A string with the five-digit U.S. FIPS code of a county
in numeric, character, or factor format.}

\item{coverage}{A numeric value in the range of 0 to 1 that specifies
the desired percentage coverage for the weather variable (i.e., what
percent of each weather variable must be non-missing to include data from
a monitor when calculating daily values averaged across monitors. The
default is to include all monitors with any available data (i.e.,
\code{coverage = 0}).}

\item{date_min}{A string with the desired starting date in character, ISO
format ("yyyy-mm-dd"). The dataframe returned will include only stations
that have data for dates including and after the specified date.}

\item{date_max}{A string with the desired ending date in character, ISO
format ("yyyy-mm-dd"). The dataframe returned will include only stations
that have data for dates up to and including the specified date.}

\item{var}{A character vector specifying desired weather variables. For
example, var = c("TMIN", "TMAX", "PRCP"). The default is \code{"all"},
which includes all available weather variables. For a full list of all
possible variable names, see NOAA's README file for the Daily Global
Historical Climatology Network (GHCN-Daily) at
\url{http://www1.ncdc.noaa.gov/pub/data/ghcn/daily/readme.txt}. Many of
the weather variables are available for some, but not all, monitors, so
your output from this function may not include all the variables
specified using this argument; if you specify a variable here but it is
not included in the output dataset, it means that it was not available in
the time range for any monitor in the county.}

\item{average_data}{TRUE / FALSE to indicate if you want the function to
average daily weather data across multiple monitors.}

\item{out_directory}{The absolute or relative pathname for the directory
where you would like the timeseries files to be saved.}

\item{out_type}{A character string indicating that you would like either .rds
files (out_type = "rds") or .csv files (out_type = ".csv"). This option
defaults to .rds files.}
}
\value{
Writes out a directory with daily weather files for each FIPS code
specified.
}
\description{
Given a vector of U.S. county FIPS codes, this function creates timeseries
dataframes giving: 1. the values for specified weather variables, and 2. the
number of weather stations contributing to the average for each day within the
specified date range.
}
\note{
If the function is unable to pull weather data for a particular county
given the specified percent coverage, date range, and/or weather variables,
\code{daily_timeseries} will not produce a file for that county.
}
\examples{
\dontrun{
daily_timeseries(fips = c("41005", "13089"), coverage = 0.90,
           date_min = "2000-01-01", date_max = "2000-01-10",
           var = c("TMAX", "TMIN", "PRCP"),
           out_directory = "~/timeseries_data")
}

}

