% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_fips.R
\name{filter_coverage}
\alias{filter_coverage}
\title{Filter stations based on "coverage" requirements}
\usage{
filter_coverage(coverage_df, percent_coverage)
}
\arguments{
\item{coverage_df}{a \code{meteo_coverage} data.frame}

\item{percent_coverage}{A numeric value in the range of 0 to 1 that specifies the
desired percentage coverage for the weather variable (i.e., what percent
of each weather variable must be non-missing to include data from a
monitor when calculating daily values averaged across monitors. The
default is 0.90 (90% non-missing observations required to include a
monitor in averaging).}
}
\value{
a \code{data.frame} with stations that meet the specified coverage
requirements for \code{prcp}, \code{tmax}, and \code{tmin}.
}
\description{
\code{filter_coverage} filters available weather variables
based on a specified required minimum coverage (i.e., percent non-missing
daily observations).
}

