% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_fips function.R
\name{fips_stations}
\alias{fips_stations}
\title{NOAA NCDC station IDs per county.}
\usage{
fips_stations(fips, date_min = NULL, date_max = NULL, data_coverage = 0)
}
\arguments{
\item{fips}{A vector of U.S. FIPS codes in numeric or factor format.}

\item{date_min}{Accepts date in character, ISO format ("yyyy-mm-dd"). The
dataframe returned will include only stations that have data for dates
including and after the specified date.}

\item{date_max}{Accepts date in character, ISO format ("yyyy-mm-dd"). The
dataframe returned will include only stations that have data for dates
including and before the specified date.}

\item{data_coverage}{A numerical value ranging from 0 to 1. The dataframe
returned will include only stations that have data coverage equal to or
greater than the specified fraction.}
}
\description{
\code{fips_stations} returns a dataframe showing NOAA NCDC station IDs for
each U.S. county present in its arguments. This function has options to
filter stations based on start and end date of available data, as well as
percent of data coverage.
}
\details{
A NOAA Token is required to use this function, which interacts with the NCDC
API. Request a Token from here: \url{http://www.ncdc.noaa.gov/cdo-web/token}.
Then run the code \code{options(noaakey = "your key")} before using this
function.
}
\examples{
\dontrun{
 fips <- c(36081, 36085, 36087, 36119, 40017)
 ex_df <- fips_stations(fips)

fips_stations(c("01073", "01089", "01097"),
   date_min = "1999-01-01", date_max = "2012-12-31",
   data_coverage = 0.9)
}

}

