% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_fips.R
\name{hourly_timeseries}
\alias{hourly_timeseries}
\title{Write hourly weather timeseries files for U.S. counties}
\usage{
hourly_timeseries(fips, coverage = NULL, year, var = "all", radius = 50,
  average_data = TRUE, out_directory)
}
\arguments{
\item{fips}{A character string or vector giving the five-digit U.S. FIPS
county code of the county or counties for which the user wants to pull
weather data.}

\item{coverage}{A numeric value in the range of 0 to 1 that specifies
the desired percentage coverage for the weather variable (i.e., what
percent of each weather variable must be non-missing to include data from
a monitor when calculating daily values averaged across monitors.)}

\item{year}{a four digit number or vector of numbers indicating the year or
years for which you want to pull hourly data. Values for \code{year} can
be in the range from 1901 to the current year.}

\item{var}{A character vector specifying desired weather variables. For
example, var = c("wind_speed", "temperature"). The core weather variables
available include \code{c("wind_direction", "wind_speed", "ceiling_height",
"visibility_distance", "temperature", "temperature_dewpoint",
"air_pressure")}. Alternatively, you can specify var = "all" to include
additional flag and quality codes.}

\item{radius}{A numeric value giving the radius, in kilometers from the
county's population-weighted center, within which to pull weather
monitors. \code{radius} is calculated from US Census Land Area data.}

\item{average_data}{TRUE / FALSE to indicate if you want the function to
average daily weather data across multiple monitors.}

\item{out_directory}{The absolute or relative pathname for the directory
where you would like the timeseries files to be saved.}
}
\value{
Writes out a directory with daily weather RDS files for each FIPS
code specified.
}
\description{
Given a vector of U.S. county FIPS codes, this function saves lists of four
elements created from the function \code{hourly_fips}. Within this list,
the element \code{hourly_data} gives a timeseries dataframe giving: 1. the
values for specified weather variables, and 2. the number of weather stations
contributing to the average for each day within the specified date range.
Other elements saved include \code{station_metadata} anmd \code{station_map}.
\code{radius}, which gives the radius (in km) within which weather stations
were pulled from each county's population-weighted center.
}
\details{
Given a vector of U.S. county FIPS codes, this function creates timeseries
dataframes giving: 1. the values for specified weather variables, and 2. the
number of weather stations contributing to the average for each day within the
specified date range.
}
\note{
If the function is unable to pull weather data for a particular county
given the specified percent coverage, date range, and/or weather variables,
\code{county_timeseries} will not produce a file for that county.
}
\examples{
\dontrun{
county_timeseries(fips = c("41005", "13089"), coverage = 0.90, year = 1992,
                  var = c("wind_speed", "temperature"),
                  out_directory = "~/timeseries_hourly")
}
}

