% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_fips.R
\name{int_surface_data}
\alias{int_surface_data}
\title{Get hourly data for a single monitor}
\usage{
int_surface_data(usaf_code, wban_code, year, var = "all")
}
\arguments{
\item{usaf_code}{A character string with a six-digit [usaf?] code for the
monitor.}

\item{wban_code}{A character string with a five-digiv [wban?] code for the
monitor.}

\item{year}{A four-digit numeric giving the year for which to pull data.}

\item{var}{A character vector listing the weather variables to pull. Choices
include ...}
}
\value{
This function returns the same type of dataframe as that returned
   by the \code{isd} function from the \code{rnoaa} package, but with the
   dataframe limited to the selected weather variables.
}
\description{
This function wraps the \code{isd} function from the \code{rnoaa} package.
}
\examples{
\dontrun{
ids <- isd_fips_stations(fips = "12086")
airport_station <- int_surface_data(usaf_code = ids$usaf[1],
                                    wban_code = ids$wban[1],
                                    year = 1992,
                                    var = c("wind_speed", "temperature"))
kendall_station <- int_surface_data(usaf_code = ids$usaf[11],
                                    wban_code = ids$wban[11],
                                    year = 1992,
                                    var = c("wind_speed", "temperature"))
}

}
\references{
For more information on this dataset, see
\url{ftp://ftp.ncdc.noaa.gov/pub/data/noaa/ish-format-document.pdf}.
}

