% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_fips.R
\name{isd_monitors_data}
\alias{isd_monitors_data}
\title{Pull hourly data for multiple monitors}
\usage{
isd_monitors_data(fips, year, var = "all")
}
\arguments{
\item{fips}{A five-digit FIPS county code.}

\item{year}{A four-digit numeric giving the year for which to pull data.}

\item{var}{A character vector listing the weather variables to pull. Choices
include ...}
}
\description{
Pull hourly data for multiple monitors
}
\examples{
\dontrun{
stationdata <- isd_monitors_data(fips = "12086", year = 1992,
                                 var = c("wind_speed", "temperature"))
}

}

