% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_fips.R
\name{mapping}
\alias{mapping}
\title{Return a dataframe with station IDs, and longitude and latitude for each
station.}
\usage{
mapping(station_df)
}
\arguments{
\item{ncdcdf}{A dataframe obtained using the \code{ncdc_stations} function
in the rnoaa package, with the \code{datasetid} argument set to 'GHCND',
the \code{locationid} set to a U.S. county FIPS code (in the format
'FIPS:08031', for example). This dataframe can be obtained using the
\code{station_fips} function. After running \code{station_fips}, the
\code{station_df} will be in your global environment.}
}
\description{
Return a dataframe with station IDs, and longitude and latitude for each
station.
}

