% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_fips.R
\name{plot_hourly_timeseries}
\alias{plot_hourly_timeseries}
\title{Write plot files for hourly weather timeseries dataframes}
\usage{
plot_hourly_timeseries(var, year, file_directory, plot_directory)
}
\arguments{
\item{var}{A character string (all lower-case) specifying which weather
variable present in the timeseries dataframe you would like to produce
plots for. For example, var = "wind_speed".}

\item{year}{A year or vecotr of years giving the year(s) present in the
timeseries dataframe.}

\item{file_directory}{The absolute or relative pathname for the directory
where your daily timeseries dataframes (produced by \code{county_timeseries})
are saved.}

\item{plot_directory}{The absolute or relative pathname for the directory
where you would like the plots to be saved.}
}
\value{
Writes out a directory with plots of timeseries data for a given
weather variable for each file present in the directory specified.
}
\description{
This function writes out a directory with plots for every timeseries file
present in the specified directory (produced by the \code{hourly_timeseries}
function) for a particular weather variable. These plots are meant to aid in
initial exploratory analysis.
}
\examples{
\dontrun{
plot_hourly_timeseries(var = "wind_speed", year = 1992,
               file_directory = "~/timeseries_hourly",
               plot_directory = "~/timeseries_plots")
}
}

