% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_fips.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Write plot files for daily weather timeseries dataframes}
\usage{
plot_timeseries(var, file_directory, file_type = "rds", plot_directory)
}
\arguments{
\item{var}{A character string (all lower-case) specifying which weather
variable present in the timeseries dataframe you would like to produce
plots for. For example, var = "prcp".}

\item{file_directory}{The absolute or relative pathname for the directory
where your daily timeseries dataframes (produced by \code{county_timeseries})
are saved.}

\item{file_type}{A character string indicating the type of timeseries files
you would like to produce plots for (either "rds" or "csv"). This option
defaults to .rds files.}

\item{plot_directory}{The absolute or relative pathname for the directory
where you would like the plots to be saved.}
}
\value{
Writes out a directory with plots of timeseries data for a given
weather variable for each file present in the directory specified.
}
\description{
This function writes out a directory with plots for every timeseries file
present in the specified directory (produced by the \code{county_timeseries}
function) for a particular weather variable. These plots are meant to aid in
initial exploratory analysis.
}
\examples{
\dontrun{
plot_timeseries(var = "prcp",
               file_directory = "~/Desktop/exposure_data/ihapps_timeseries",
               plot_directory = "~/Desktop/exposure_data/plots_prcp")

plot_timeseries(files = files, var = "tmax",
               file_directory = "~/Desktop/exposure_data/ihapps_timeseries",
               plot_directory = "~/Desktop/exposure_data/plots_tmax")

plot_timeseries(var = "tmin",
               file_directory = "~/Desktop/exposure_data/ihapps_timeseries",
               plot_directory = "~/Desktop/exposure_data/plots_tmin")
}
}

