% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_fips.R
\name{station_radius}
\alias{station_radius}
\title{Search for stations located within a specified radius for a particular
U.S. county.}
\usage{
station_radius(fips, radius = NULL)
}
\value{
A character vector listing station ids which are located within
   the specified radius centered in the specified county.
}
\description{
Search for stations located within a specified radius for a particular
U.S. county.
}

