% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationspercounty function.R
\name{stations_per_county}
\alias{stations_per_county}
\title{Count of weather stations per county.}
\usage{
stations_per_county(yourvector)
}
\arguments{
\item{yourvector}{A vector of U.S. FIPS codes in numberic or factor format.}
}
\description{
\code{stations_per_county} returns a plot showing the number of GHCND weather
stations for each U.S. county present in its arguments.
}
\details{
A NOAA Token is required to use this function, which interacts with the NCDC
API. Request a Token from here: \url{http://www.ncdc.noaa.gov/cdo-web/token}.
Then run the code \code{options(noaakey = "your key")} before using this
function.
}
\examples{
vec7 <- c(36081, 36085, 36087, 36119, 40017)
stations_per_county(vec7)

returns an error - none of these fips had relevant stations
\donttest{
vec2 <- c("01073", "01089", "01097", "01101", "02020", "04013", "04019",
         "05119", "06001", "06013", "06019", "06029", "06037", "06065",
         "06067", "06071", "06073", "06075", "06077", "06081", "06085",
         "06099", "06111", "08001", "08005", "08031", "08041", "09001",
         "09003", "09009", "10003", "11001", "12001", "12011", "12031",
         "12033", "12057", "12071", "12073", "12081", "12086", "12095",
         "12099", "12103", "12117", "12127", "13021", "13051", "13063",
         "13067", "13089", "13121", "13215", "13245")
stations_per_county(vec2)
}

}

