% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Floods.R
\name{stream_data}
\alias{stream_data}
\title{Return average daily streamflow data for a particular county and date range.}
\usage{
stream_data(fips, date_min, date_max, stat_code = "00003")
}
\arguments{
\item{fips}{A character string or character vector giving the five-digit
U.S. FIPS county code(s) of the county for which the user wants to pull
hydrologic data.}

\item{date_min}{A character string giving the earliest date you want to
pull data for in "yyyy-mm-dd" format.}

\item{date_max}{A character string giving the latest date you want to pull
data for in "yyyy-mm-dd" format.}

\item{stat_code}{A 5-number character string specifying a USGS statistics
code to uniquely identify specific statistics. Possible codes can be found
here:
\url{http://help.waterdata.usgs.gov/code/stat_cd_nm_query?stat_nm_cd=\%25&fmt=html}.
Not all statistics are available at every gage. The default for stat_code in
\code{stream_data} is set to "00003", which gives daily mean values.}
}
\description{
\code{stream_data} returns a data.frame showing USGS station IDs, mean daily
streamflow values, dates, and a quantification code. This function applies
the \code{importDVs} function from the waterData package for all of the USGS
stations for a given FIPS code.
}
\examples{
ex <- stream_data("47157", "2011-01-01", "2011-12-31")

}

