% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Floods.R
\name{stream_stations}
\alias{stream_stations}
\title{USGS Daily Streamflow data per US counties}
\usage{
stream_stations(fips, date_min, date_max, fraction_coverage = NULL)
}
\arguments{
\item{fips}{A character string or character vector giving the five-digit
U.S. FIPS county code(s) of the county for which the user wants to pull
hydrologic data.}

\item{date_min}{A character string giving the earliest date you want to
pull data for in "yyyy-mm-dd" format.}

\item{date_max}{A character string giving the latest date you want to pull
data for in "yyyy-mm-dd" format.}

\item{fraction_coverage}{A numeric value between 0 and 1. The dataframe
returned will include only stations that have data coverage equal to or
greater than the specified fraction. (Optional.)}
}
\description{
\code{stream_stations} returns a vector of U.S. Geological Survey (USGS)
station identification numbers for a given US county, date range, and percent
coverage.
}
\details{
This function interacts with the USGS Site web service to import daily
hydrological data. Options at the USGS Site Web Service URL Generation Tool
(\url{http://waterservices.usgs.gov/rest/Site-Test-Tool.html}) are set to
the following settings:
\itemize{
 \item Major Filters = Counties
 \item Optional Arguments: Show period of record information about these
 data types = Daily values
 \item Optional Filters = Show only sites operational between a start date
 and an end date
 \item Optional Filters: Show only sites serving these data types = Daily
 values
 \item Optional Filters: Sites serving parameter codes = 00060
}
}
\examples{
ex <- stream_stations("47157", "2011-01-01", "2011-12-31")

}

