% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Floods.R
\name{streamflow}
\alias{streamflow}
\title{USGS Daily Streamflow data per US counties}
\usage{
streamflow(fips, date_min, date_max)
}
\description{
\code{}
}
\examples{
fipsvec <- c(51059, 51061)
test <- streamflow(fipsvec, "2000-04-12", "2005-04-13")
}

