% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_fips function.R, R/weather_fips new.R
\name{weather_fips}
\alias{weather_fips}
\title{Daily precipiation, maximum and minimum temperatures per county.}
\usage{
weather_fips(fips, coverage_val, min_date, max_date)

weather_fips(fips, coverage_val, min_date, max_date)
}
\arguments{
\item{fips}{a U.S. FIPS code in character format}

\item{coverage_val}{a number in the range of 0 to 1 that specifies the desired
coverage.}

\item{station_fips_df}{A dataframe containing NCDC station IDs in character
format ("USC00010764", for example) and corresponding fips codes in factor
format. This dataframe coulld be obtained for a given vector of fips codes
with the \code{fips_stations} function.}

\item{start_date}{a date in "yyyy-mm-dd" format - the earliest date you want
in your dataset.
\code{start_date}.}

\item{end_date}{a date in "yyyy-mm-dd" format - the lastest date you want in
your dataset.}
}
\description{
\code{weather_fips} returns a dataframe showing GHCN-Daily weather data
(precipitation in mm, and maximum and minumum temperatures in fahrenheit and
celsius), given a dataframe of NOAA NCDC station IDs and corresponding U.S.
county fips codes. (This dataframe could be obtained with the
\code{fips_stations} function.)

\code{weather_fips} returns a data.frame of average daily precipitation,
maximum and minimum temperature values for a particular county, date range,
and specified "coverage."
}
\details{
A NOAA Token is required to use this function, which interacts with the NCDC
API. Request a Token from here: \url{http://www.ncdc.noaa.gov/cdo-web/token}.
Then run the code \code{options(noaakey = "your key")} before using this
function.
}
\examples{
\dontrun{
ex_df <- fips_stations(c("01073", "01089", "01097"),
         date_min = "1999-01-01",
         date_max = "2012-12-31", data_coverage = 0.9)
weather_data <- weather_fips(ex_df)
head(weather_data)
}
\dontrun{
df <- weather_fips("06037", 0.90, "1999-01-01", "2012-12-31")
}
}

