% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_fips.R
\name{weather_fips_df}
\alias{weather_fips_df}
\title{Return average daily weather data for a particular county.}
\usage{
weather_fips_df(fips, percent_coverage = NULL, date_min = NULL,
  date_max = NULL, var = "all")
}
\arguments{
\item{fips}{A character string giving the five-digit U.S. FIPS county code
of the county for which the user wants to pull weather data.}

\item{percent_coverage}{A numeric value in the range of 0 to 1 that specifies
the desired percentage coverage for the weather variable (i.e., what
percent of each weather variable must be non-missing to include data from
a monitor when calculating daily values averaged across monitors.
(Optional.)}

\item{date_min}{A character string giving the earliest date you want
   in your dataset in "yyyy-mm-dd" format. (Optional.)
\code{date_min}.}

\item{date_max}{A character string giving the latest date you want
in your dataset in "yyyy-mm-dd" format. (Optional.)}

\item{var}{A character vector specifying desired weather variables. For
example, var = c("TMIN", "TMAX", "PRCP"). (Optional.)}
}
\value{
A dataframe of daily weather data averaged across multiple monitors,
   as well as columns (\code{"var"_reporting}) for each weather variable
   showing the number of stations contributing to the average for that
   variable on that day.
}
\description{
\code{weather_fips_df} returns a dataframe of average daily weather values
for a particular county, date range, and/or specified "coverage."
}
\details{
This function serves as a wrapper to several functions from the \code{rnoaa}
package, which provides weather data from all relevant stations in a county.
This function filters and averages across stations based on user-specified
coverage specifications.
}
\note{
Because this function uses the NOAA API to identify the weather
   monitors within a US county, you will need to get an access token from
   NOAA to use this function. Visit NOAA's token request page
   (\url{http://www.ncdc.noaa.gov/cdo-web/token}) to request a token by
   email, and then use the code
   \code{options("noaakey" = "<key NOAA emails you>")} to set up your
   API access.
}
\examples{
\dontrun{
df <- weather_fips_df(fips = "12086",
                   percent_coverage = 0.90, date_min = "2010-01-01",
                   date_max = "2010-02-01", var = c("TMAX", "TMIN", "PRCP"))
}
}

