% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_helpers.R
\name{ave_hourly}
\alias{ave_hourly}
\title{Average hourly weather data across multiple stations.}
\usage{
ave_hourly(hourly_data)
}
\arguments{
\item{hourly_data}{A dataframe with hourly weather observations. This
dataframe is returned from the function \code{isd_monitors_data}.}
}
\description{
\code{ave_hourly} returns a dataframe with hourly weather averaged across
stations, as well as columns showing the number of stations
}
\details{
\code{ave_hourly} returns a dataframe with hourly weather averaged across
stations, as well as columns showing the number of stations contributing to
average for each variable and each hour.
}

