% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_fips.R
\name{daily_fips}
\alias{daily_fips}
\title{Return average daily weather data and a plot showing the location of weather
stations for a particular county.}
\usage{
daily_fips(fips, coverage = NULL, date_min = NULL, date_max = NULL,
  var = "all", average_data = TRUE, station_label = FALSE)
}
\arguments{
\item{coverage}{A numeric value in the range of 0 to 1 that specifies
the desired percentage coverage for the weather variable (i.e., what
percent of each weather variable must be non-missing to include data from
a monitor when calculating daily values averaged across monitors.
(Optional.)}

\item{date_min}{A character string giving the earliest date you want
   in your dataset in "yyyy-mm-dd" format. (Optional.)
\code{date_min}.}

\item{date_max}{A character string giving the latest date you want
in your dataset in "yyyy-mm-dd" format. (Optional.)}

\item{var}{A character vector specifying desired weather variables. For
example, var = c("TMIN", "TMAX", "PRCP"). (Optional.)}

\item{average_data}{TRUE / FALSE to indicate if you want the function to
average daily weather data across multiple monitors.}

\item{station_label}{TRUE / FALSE to indicate if you want your plot of
weather station locations to include labels indicating station ids.}
}
\value{
A list with two elements. The first element (\code{daily_data}) is a
   dataframe of daily weather data averaged across multiple stations, as well
   as columns (\code{"var"_reporting}) for each weather variable showing the
   number of stations contributing to the average for that variable on that
   day. The second element (\code{station_map}) is a plot showing points for all
   weather stations for a particular county satisfying the conditions present
   in \code{daily_fips}'s arguments (coverage, date_min, date_max,
   and/or var).
}
\description{
Given a particular county FIPS code, this function returns a list with two
elements: "data", a dataframe of daily average weather values, and "plot",
a plot showing the location of weather stations contributing to the average
weather in "data".
}
\note{
You must have a NOAA API to use this function, and you need to set
   that API code in your R session (e.g., using
   \code{options(noaakey = "your key"}).
}
\examples{
\dontrun{
ex <- daily_fips("08031", coverage = 0.90,
                   date_min = "2010-01-01", date_max = "2010-02-01",
                   var = "PRCP")

weather_data <- ex$daily_data
station_map <- ex$station_map

mobile_ex <- weather_fips("01097", percent_coverage = 0,
                          date_min = "1997-07-13",
                          date_max = "1997-07-25",
                          var = "PRCP", average_data = FALSE)
library(ggplot2)
ggplot(mobile_ex$daily_weather, aes(x = date, y = prcp, color = id)) +
       geom_line()
}
}

