% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_helpers.R
\name{filter_coverage}
\alias{filter_coverage}
\title{Filter stations based on "coverage" requirements.}
\usage{
filter_coverage(coverage_df, coverage = NULL)
}
\arguments{
\item{coverage_df}{A \code{meteo_coverage} dataframe}

\item{coverage}{A numeric value in the range of 0 to 1 that specifies
the desired percentage coverage for the weather variable (i.e., what
percent of each weather variable must be non-missing to include data from
a monitor when calculating daily values averaged across monitors.
(Optional.)}
}
\value{
a \code{dataframe} with stations that meet the specified coverage
   requirements for weather variables included in the dataframe present in
   this function's arguments.
}
\description{
\code{filter_coverage} filters available weather variables
based on a specified required minimum coverage (i.e., percent non-missing
daily observations).
}

