% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_fips new.R
\name{filter_coverage}
\alias{filter_coverage}
\title{Filter stations based on "coverage" requirements}
\usage{
filter_coverage(coverage_df, percent_coverage)
}
\arguments{
\item{coverage_df}{a \code{meteo_coverage} data.frame}

\item{percent_coverage}{a number ranging from 0 to 1. Resulting coverage for
precipitation, maximum and minimum temperatures will be equal to or greater
than the specified value.}
}
\value{
a \code{data.frame} with stations that meet the specified coverage
requirements for \code{prcp}, \code{tmax}, and \code{tmin}.
}
\description{
\code{filter_coverage} filters precipitation, maximum and minumum temperature
based on a specified value of coverage.
}

