% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_helpers.R
\name{filter_hourly}
\alias{filter_hourly}
\title{Filter NOAA ISD stations based on "coverage" requirements.}
\usage{
filter_hourly(hourly_data, coverage, var = "all")
}
\arguments{
\item{hourly_data}{A \code{isd_monitors_data} dataframe (The "df" element of
a \code{isd_monitors_data} list)}

\item{coverage}{A numeric value in the range of 0 to 1 that specifies the
desired percentage coverage for each weather variable (i.e., what percent
of each weather variable must be non-missing to include the data from a
station when calculating hourly values averaged across stations).}

\item{var}{A character vector specifying desired weather variables. For
example, var = c("wind_speed", "temperature"). The core avaialbe weather
variables include \code{c("wind_direction", "wind_speed", "ceiling_height",
"visibility_distance", "temperature", "temperature_dewpoint",
"air_pressure")}. Alternatively, you can specify var = "all" to include
additional flag and quality codes.}
}
\value{
A list with two elements: \code{stations} is a vector of station ids
   (usaf and wban identification numbers pasted together, separated by "-")
   that meet the specified coverage requirements. \code{df} is a dataframe
   with weather data from stations that meet the coverage requirements.
}
\description{
\code{filter_hourly} filters available weather variables based on a specified
minimum coverage (i.e., percent non-missing hourly observations).
}

