% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_fips.R
\name{filter_hourly}
\alias{filter_hourly}
\title{Filter NOAA ISD stations based on "coverage" requirements.}
\usage{
filter_hourly(hourly_data, coverage, var)
}
\arguments{
\item{hourly_data}{A \code{isd_monitors_data} dataframe}

\item{coverage}{A numeric value in the range of 0 to 1 that specifies the
desired percentage coverage for each weather variable (i.e., what percent
of each weather variable must be non-missing to include the data from a
station when calculating hourly values averaged across stations). (Optional).}
}
\value{
a \code{dataframe} with stations that meet the specified coverage
requirements for weather variables included in the datafrome present in
this function's arguments.
}
\description{
\code{filter_hourly} filters available weather variables based on a specified
minimum coverage (i.e., percent non-missing hourly observations).
}

