% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_fips.R
\name{hourly_df}
\alias{hourly_df}
\title{Return average hourly weather data for a particular county.}
\usage{
hourly_df(fips, year, var = "all", average_data = TRUE, coverage = NULL)
}
\arguments{
\item{fips}{A character string or vector giving the five-digit U.S. FIPS
county code of the county or counties for which the user wants to pull
weather data.}

\item{year}{a four digit number or vector of numbers indicating the year or
years for which you want to pull hourly data. Values for \code{year} can
be in the range from 1901 to the current year.}

\item{var}{A character vector specifying desired weather variables. For
example, var = c("wind_speed", "temperature"). The core weather variables
available include \code{c("wind_direction", "wind_speed", "ceiling_height",
"visibility_distance", "temperature", "temperature_dewpoint",
"air_pressure")}. Alternatively, you can specify var = "all" to include
additional flag and quality codes.}

\item{average_data}{TRUE / FALSE to indicate if you want the function to
average daily weather data across multiple monitors.}

\item{coverage}{A numeric value in the range of 0 to 1 that specifies
the desired percentage coverage for the weather variable (i.e., what
percent of each weather variable must be non-missing to include data from
a monitor when calculating daily values averaged across monitors.)}

\item{radius}{A numeric value giving the radius, in kilometers from the
county's population-weighted center, within which to pull weather
monitors. \code{radius} is calculated from US Census Land Area data.}
}
\value{
A list with three elements. The first element, \code{hourly_data}, is
   a dataframe of hourly weather data averaged across multiple stations,
   as well as columns (\code{"var"_reporting}) for each weather variable
   showing the number of stations contributing to the average for that
   variable for each hour. \code{station_df} is a dataframe of station
   metadata for each station contributing weather data. \code{radius} is the
   calculated radius within which stations were pulled from the county's
   population-weighted center.
}
\description{
\code{hourly_df} returns a dataframe of average daily weather values
for a particular county, year, and/or specified "coverage."
}
\details{
This function serves as a wrapper to several functions from the \code{rnoaa}
package, which provides weather data from all relevant stations in a county.
This function filters and averages across NOAA ISD/ISH stations based on
user-specified coverage specifications.
}
\note{
: Observation times are vased on Coordinated Universal Time Code (UTC).
}
\examples{
\dontrun{
df <- hourly_df(fips = "12086", year = 1992, var = c("wind_speed",
                     "temperature"))
data <- df$hourly_data
station_info <- df$station_df
radius <- df$radius
}

}
\references{
For more information on this dataset and available weather and
flag/quality variables, see
\url{ftp://ftp.ncdc.noaa.gov/pub/data/noaa/ish-format-document.pdf}.
}

