% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_fips.R
\name{hourly_stationmap}
\alias{hourly_stationmap}
\title{Plot hourly weather stations for a particular county}
\usage{
hourly_stationmap(fips, hourly_data, point_color = "firebrick",
  point_size = 2)
}
\arguments{
\item{point_size}{}
}
\description{
Plot hourly weather stations for a particular county
}
\examples{
\dontrun{
hourly_data <- hourly_fips_df <- (fips = "12086", year = 1992,
                                  var = c("wind_speed", "temperature"))
hourly_stationmap("12086", hourly_data)
}

}

