% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_helpers.R
\name{hourly_stationmap}
\alias{hourly_stationmap}
\title{Plot hourly weather stations for a particular county}
\usage{
hourly_stationmap(fips, hourly_data, point_color = "firebrick",
  point_size = 2, station_label = FALSE)
}
\arguments{
\item{fips}{A five-digit FIPS county code.}

\item{hourly_data}{hourly_data A dataframe with hourly weather observations. This
dataframe is returned from the function \code{isd_monitors_data}.}

\item{point_color}{point_color The specified \code{ggplot2} color for each point
representing the location of a station.}

\item{point_size}{The specified \code{ggplot2} size for each point
representing the location of a station.}

\item{station_label}{TRUE / FALSE If TRUE, includes labels giving the id for
monitor on the map. The default is FALSE.}
}
\value{
A plot showing points for all weather stations for a particular
   county satisfying the conditions present in \code{hourly_df}'s
   arguments (year(s) and/or var).
   (\code{stationmap_fips} takes the resulting weather dataframe from this
   function.)
}
\description{
Plot hourly weather stations for a particular county
}
\examples{
\dontrun{
hourly_data <- hourly_df(fips = "12086", year = 1992,
                         var = c("wind_speed", "temperature"))
hourly_stationmap("12086", hourly_data)
}

}

