% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_fips.R
\name{isd_fips_stations}
\alias{isd_fips_stations}
\title{Get station list for a particular fips}
\usage{
isd_fips_stations(fips)
}
\arguments{
\item{fips}{A five-digit FIPS county code.}
}
\value{
A dataframe of monitors within a given radius of the
   population-weighted center of the county specified by the FIPS code.
   This will have the same dataframe format as the output from the
   \code{isd_stations_search} function in the \code{rnoaa} package.
}
\description{
This function serves as a wrapper to that function, allowing you to search
by FIPS code rather than having to know the latitude and longitude of the
center of each county.
}
\note{
We probably want to use geocodes for this instead.
}
\examples{
\dontrun{
ids <- isd_fips_stations(fips = "12086")
}

}

