% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_fips.R
\name{isd_fips_stations}
\alias{isd_fips_stations}
\title{Get station list for a particular US county}
\usage{
isd_fips_stations(fips, verbose = TRUE, radius = 50)
}
\arguments{
\item{fips}{A five-digit FIPS county code.}

\item{verbose}{TRUE / FALSE to indicate if you want the function to print
out the name of the county it's processing}

\item{radius}{A numeric value giving the radius, in kilometers from the
county's population-weighted center, within which to pull weather
monitors.}
}
\value{
A dataframe of monitors within the given radius of the
   population-weighted center of the county specified by the FIPS code.
   This will have the same dataframe format as the output from the
   \code{isd_stations_search} function in the \code{rnoaa} package.
}
\description{
This function serves as a wrapper to the \code{isd_stations_search} function
in the \code{rnoaa} package, allowing you to search by FIPS code rather than
having to know the latitude and longitude of the center of each county.
}
\examples{
\dontrun{
ids <- isd_fips_stations(fips = "12086")
}

}

