% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_fips.R
\name{isd_monitors_data}
\alias{isd_monitors_data}
\title{Pull hourly data for multiple monitors}
\usage{
isd_monitors_data(fips, year, var = c("wind_direction", "wind_speed",
  "ceiling_height", "visibility_distance", "temperature",
  "temperature_dewpoint", "air_pressure"), radius = 50)
}
\arguments{
\item{fips}{A five-digit FIPS county code.}

\item{year}{A four-digit numeric giving the year for which to pull data.}

\item{var}{A character vector listing the weather variables to pull. Choices
include ...}

\item{radius}{A numeric value giving the radius, in kilometers from the
county's population-weighted center, within which to pull weather
monitors.}
}
\description{
Pull all available data for all weather monitors within a certain radius of
the population-weighted center of a US county, based on the county's FIPS
code.
}
\examples{
\dontrun{
stationdata <- isd_monitors_data(fips = "12086", year = 1992,
                                 var = c("wind_speed", "temperature"))
ggplot(stationdata, aes(x = date_time, y = wind_speed)) +
   geom_point(alpha = 0.5, size = 0.2) +
   facet_wrap(~ usaf_station, ncol = 1)
}
}

