% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_fips.R
\name{plot_daily_timeseries}
\alias{plot_daily_timeseries}
\title{Write plot files for daily weather timeseries dataframes}
\usage{
plot_daily_timeseries(var, file_directory, plot_directory, date_min, date_max)
}
\arguments{
\item{var}{A character string (all lower-case) specifying which weather
variable present in the timeseries dataframe you would like to produce
plots for. For example, var = "prcp".}

\item{file_directory}{The absolute or relative pathname for the directory
where your daily timeseries dataframes (produced by \code{daily_timeseries})
are saved.}

\item{plot_directory}{The absolute or relative pathname for the directory
where you would like the plots to be saved.}

\item{date_min}{A character string giving the earliest date present in the
timeseries dataframe in "yyyy-mm-dd" format.}

\item{date_max}{A character string giving the latest date present in the
timeseries dataframe in "yyyy-mm-dd" format.}
}
\value{
Writes out a directory with plots of timeseries data for a given
weather variable for each file present in the directory specified.
}
\description{
This function writes out a directory with plots for every timeseries file
present in the specified directory (produced by the \code{daily_timeseries}
function) for a particular weather variable. These plots are meant to aid in
initial exploratory analysis.
}
\examples{
\dontrun{
plot_daily_timeseries(var = "prcp",
               file_directory = "~/Desktop/exposure_data/ihapps_timeseries",
               plot_directory = "~/Desktop/exposure_data/plots_prcp")

plot_daily_timeseries(files = files, var = "tmax",
               file_directory = "~/Desktop/exposure_data/ihapps_timeseries",
               plot_directory = "~/Desktop/exposure_data/plots_tmax")

plot_daily_timeseries(var = "tmin",
               file_directory = "~/Desktop/exposure_data/ihapps_timeseries",
               plot_directory = "~/Desktop/exposure_data/plots_tmin")
}
}

