% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_fips.R
\name{stationmap_fips}
\alias{stationmap_fips}
\title{Plot weather stations for a particular county}
\usage{
stationmap_fips(fips, weather_data, point_color = "firebrick",
  point_size = 2)
}
\arguments{
\item{fips}{A character string giving the five-digit U.S. FIPS county code
of the county for which the user wants to pull weather data.}
}
\value{
A plot showing points for all weather stations for a particular
   county satisfying the conditions present in \code{stationmap_fips}'s
   arguments (percent_coverage, date_min, date_max, and/or var).
}
\description{
Plot weather stations for a particular county
}
\examples{
\dontrun{
stations <- fips_stations(fips = "12086", date_min = "1999-08-01",
                          date_max = "1999-08-31")
weather_data <- weather_fips_df(stations = stations, percent_coverage = 0.90,
                      var = c("PRCP"), date_min = "1999-08-01",
                          date_max = "1999-08-31")
stationmap_fips(fips = "12086", weather_data = weather_data)
}

}

