% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_fips.R
\name{stationmap_fips}
\alias{stationmap_fips}
\title{Plot of stations for a particular FIPS}
\usage{
stationmap_fips(fips, percent_coverage, date_min, date_max)
}
\description{
Plot of stations for a particular FIPS
}
\examples{
\dontrun{
ex <- stationmap_fips("08031", 0.90, "2000-01-01", "2010-12-31")
}
}

