% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Floods.R
\name{stream_floods}
\alias{stream_floods}
\title{Return average daily streamflow data for a particular county and date range.}
\usage{
stream_floods(fips, date_min, date_max, stat_code)
}
\description{
Return average daily streamflow data for a particular county and date range.
}

