% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_fips function.R
\name{weather_fips}
\alias{weather_fips}
\title{Daily precipiation, maximum and minimum temperatures per county.}
\usage{
weather_fips(station_fips_df)
}
\arguments{
\item{station_fips_df}{A dataframe containing NCDC station IDs in character
format ("USC00010764", for example) and corresponding fips codes in factor
format. This dataframe coulld be obtained for a given vector of fips codes
with the \code{fips_stations} function.}
}
\description{
\code{weather_fips} returns a dataframe showing GHCN-Daily weather data
(precipitation in mm, and maximum and minumum temperatures in fahrenheit and
celsius), given a dataframe of NOAA NCDC station IDs and corresponding U.S.
county fips codes. (This dataframe could be obtained with the
\code{fips_stations} function.)
}
\details{
A NOAA Token is required to use this function, which interacts with the NCDC
API. Request a Token from here: \url{http://www.ncdc.noaa.gov/cdo-web/token}.
Then run the code \code{options(noaakey = "your key")} before using this
function.
}
\examples{
\dontrun{
ex_df <- fips_stationsc("01073", "01089", "01097"), datemin = "1999-01-01",
         datemax = "2012-12-31", datacov = 0.9)
weather_data <- weather_fips(ex_df)
head(weather_data)
}
}

